# docs/api/swagger.yaml
openapi: 3.0.0
info:
  title: Fuel Pre-Order API
  description: API for fuel pre-order and pickup mobile application
  version: 1.0.0
  contact:
    name: Fuel App Team
    email: support@fuelapp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.fuelapp.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /orders:
    post:
      tags:
        - Orders
      summary: Create new fuel order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - stationId
                - fuelType
                - quantity
                - scheduledTime
              properties:
                stationId:
                  type: string
                fuelType:
                  type: string
                  enum: [gasoline, diesel, electric, lpg, cng, premium_gasoline, premium_diesel]
                quantity:
                  type: number
                unit:
                  type: string
                  enum: [liters, gallons, kWh]
                  default: liters
                scheduledTime:
                  type: string
                  format: date-time
                vehicle:
                  type: object
                addOns:
                  type: array
                  items:
                    type: object
      responses:
        201:
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      order:
                        $ref: '#/components/schemas/Order'

    get:
      tags:
        - Orders
      summary: Get user orders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, preparing, ready, in_progress, completed, cancelled, expired]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      orders:
                        type: array
                        items:
                          $ref: '#/components/schemas/Order'
                      pagination:
                        type: object
                        properties:
                          total:
                            type: integer
                          limit:
                            type: integer
                          skip:
                            type: integer
                          hasMore:
                            type: boolean
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Order details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      order:
                        $ref: '#/components/schemas/Order'
  /orders/{id}/cancel:
    put:
      tags:
        - Orders
      summary: Cancel order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        200:
          description: Order cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      order:
                        $ref: '#/components/schemas/Order'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        user:
          type: string
        station:
          type: string
        fuelType:
          type: string
        quantity:
          type: number
        unit:
          type: string
        pricing:
          type: object
          properties:
            basePrice:
              type: number
            fuelPrice:
              type: number
            tax:
              type: number
            serviceFee:
              type: number
            totalAmount:
              type: number
            currency:
              type: string
        pickup:
          type: object
          properties:
            scheduledTime:
              type: string
              format: date-time
            actualTime:
              type: string
              format: date-time
            status:
              type: string
            qrCode:
              type: string
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, in_progress, completed, cancelled, expired]
        createdAt:
          type: string
          format: date-time
...